save_history/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.6